Preliminary 1.18.2 support
This allows the mod to build/load/run. However some further changes are desirable.

Most of this commit is a rename on ColorAlpha --> alpha (and equivalents for red/green/blue)

Code for the biome type changes in ClientWorldMixin and FireflyParticle will need additional work in future as biome.getCategory() has been deprecated.